% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acp.r
\name{acp}
\alias{acp}
\title{Principal Component Analysis}
\usage{
acp(
  dat,
  factor = NULL,
  samplename = NULL,
  pc1 = 1,
  pc2 = 2,
  center = TRUE,
  scale = TRUE,
  title = "ACP",
  legendtitle = "TREATMENT"
)
}
\arguments{
\item{dat}{matrix numeric}

\item{factor}{factor}

\item{samplename}{character}

\item{pc1}{numeric}

\item{pc2}{numeric}

\item{center}{logical TRUE}

\item{scale}{logical TRUE}

\item{title}{character}

\item{legendtitle}{character}
}
\value{
no values
}
\description{
Principal Component Analysis
}
\examples{
# not run
# acp( mat1 , sif1 )
}
\author{
Florent Dumont \href{mailto:florent.dumont@univresite-paris-saclay.fr}{florent.dumont@univresite-paris-saclay.fr}
}
