% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.r
\name{input}
\alias{input}
\title{import tab file in data.frame}
\usage{
input(filename, sep = "\\t", quote = "")
}
\arguments{
\item{filename}{character path to the file to read}

\item{sep}{character for field separator}

\item{quote}{character for field quote}
}
\value{
data.frame
}
\description{
import tab file in data.frame
}
\details{
wrapper of read.table function for tabular separated files
}
\examples{
# not run
# input( "filename" ) -> dt
}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
