% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.r
\name{output}
\alias{output}
\title{export data.frame in tab file}
\usage{
output(dt, filename)
}
\arguments{
\item{dt}{data.frame}

\item{filename}{character}
}
\description{
export data.frame in tab file
}
\examples{
# not run
# output( dt )
}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
